update variable types to match those found in dependency

+++ internal/cmd/cli/gitcloner/cmd.go
@@ -18,8 +18,8 @@ type GitCloner struct {
 	PasswordFile          string
 	SSHPrivateKeyFile     string
 	InsecureSkipTLS       bool
-	GitHubAppID           int
-	GitHubAppInstallation int
+	GitHubAppID           int64
+	GitHubAppInstallation int64
 	GitHubAppKeyFile      string
 }
 
@@ -45,8 +45,8 @@ func NewCmd(gitCloner CloneGit) *cobra.Command {
 	cmd.Flags().StringVar(&opts.PasswordFile, "password-file", "", "password file for basic auth")
 	cmd.Flags().StringVar(&opts.SSHPrivateKeyFile, "ssh-private-key-file", "", "ssh private key file path")
 	cmd.Flags().BoolVar(&opts.InsecureSkipTLS, "insecure-skip-tls", false, "do not verify tls certificates")
-	cmd.Flags().IntVar(&opts.GitHubAppID, "github-app-id", 0, "GitHub App ID")
-	cmd.Flags().IntVar(&opts.GitHubAppInstallation, "github-app-installation-id", 0, "GitHub App installation ID")
+	cmd.Flags().Int64Var(&opts.GitHubAppID, "github-app-id", 0, "GitHub App ID")
+	cmd.Flags().Int64Var(&opts.GitHubAppInstallation, "github-app-installation-id", 0, "GitHub App installation ID")
 	cmd.Flags().StringVar(&opts.GitHubAppKeyFile, "github-app-key-file", "", "path to GitHub App private-key PEM")
 
 	return cmd
